Update contact form based on field mapping
Story https://app.shortcut.com/deskpro/story/165035/hubspot-contact-field-mapping
