--- src/components/ContactForm/index.ts
@@ -1,1 +1,0 @@
-export { ContactForm } from "./ContactForm";
