+++ src/components/UpdateContact/index.ts
@@ -1,0 +1,1 @@
+export { UpdateContact } from "./UpdateContact";
