+++ src/components/common/Builder/BlocksBuilder/GenerateBlock.tsx
@@ -9,26 +9,9 @@ type Props = {
 };
 
 const GenerateBlock = ({ meta, Component, value }: Props) => {
-  const blockType = meta?.type;
-  const label = meta?.label;
-
-  if (!meta) {
-    // eslint-disable-next-line no-console
-    console.error("BlocksBuilder: wrong config - block config not found");
-    return null;
-  }
-
-  if (!Component) {
-    // eslint-disable-next-line no-console
-    console.error(
-      "BlocksBuilder: can't find component for block type:",
-      blockType,
-    );
-  }
-
   return (
     <Property
-      label={label}
+      label={meta.label}
       marginBottom={0}
       text={Component && (
         <Component meta={meta} value={value} />
