+++ src/components/common/Builder/FormBuilder/index.ts
@@ -1,2 +1,2 @@
 export { FormBuilder } from "./FormBuilder";
-export type { FieldProps, FormBuilderProps } from "./types";
+export type { FieldProps, FormValues, FormBuilderProps } from "./types";
