+++ src/components/common/Builder/FormBuilder/types.ts
@@ -10,7 +10,7 @@ import type { Layout } from "../types";
 
 export type Meta = Record<PropertyMeta["name"], PropertyMeta>;
 
-export type Values = Record<PropertyMeta["name"], string>;
+export type FormValues = Record<PropertyMeta["name"], string>;
 
 export type Config = {
     structure: Layout;
@@ -22,9 +22,10 @@ export type FieldsMap = Record<PropertyMeta["fieldType"], FC<FieldProps>>;
 export type FormBuilderProps = {
     config: Config;
     fieldsMap: FieldsMap;
-    onSubmit: SubmitHandler<Values>;
-    values?: Values;
+    onSubmit: SubmitHandler<FormValues>;
+    values?: FormValues;
     onCancel?: () => void;
+    isEditMode?: boolean;
 };
 
 export type FieldProps<T = unknown> = {
