+++ src/components/common/Builder/index.ts
@@ -3,4 +3,4 @@ export { BlocksBuilder } from "./BlocksBuilder";
 export { FormBuilder } from "./FormBuilder";
 export type { Layout } from "./types";
 export type { BlockProps, BlocksBuilderProps } from "./BlocksBuilder";
-export type { FieldProps, FormBuilderProps } from "./FormBuilder";
+export type { FieldProps, FormValues, FormBuilderProps } from "./FormBuilder";
