+++ src/components/index.ts
@@ -1,6 +1,6 @@
-export { ContactForm } from "./ContactForm";
 export { DealForm } from "./DealForm";
 export { NoteForm } from "./NoteForm";
 export { ViewContact } from "./ViewContact";
 export { CreateContact } from "./CreateContact";
+export { UpdateContact } from "./UpdateContact";
 export * from "./Admin";
