+++ src/hooks/index.ts
@@ -2,7 +2,6 @@ export { useQueryWithClient } from "./useQueryWithClient";
 export { useSetAppTitle } from "./useSetAppTitle";
 export { useCheckLinkedContact } from "./useCheckLinkedContact";
 export { useQueriesWithClient } from "./useQueriesWithClient";
-export { useLoadUpdateContactDeps } from "./useLoadUpdateContactDeps";
 export { useLoadUpdateDealDeps } from "./useLoadUpdateDealDeps";
 export { useOwnerOptions } from "./useOwnerOptions";
 export { useLinkUnlinkNote } from "./useLinkUnlinkNote";
@@ -10,3 +9,4 @@ export { useLinkContact } from "./useLinkContact";
 export { useUnlinkContact } from "./useUnlinkContact";
 export { useAppContext, AppProvider } from "./AppContext";
 export { useAsyncError } from "./useAsyncError";
+export { useContact } from "./useContact";
