+++ src/utils/getFullName.ts
@@ -1,7 +1,9 @@
 import { isContact, isOwner } from "../utils";
 import type { Contact, Owner } from "../services/hubspot/types";
 
-type User = Partial<Contact["properties"] | Owner>;
+type User =
+  | Pick<Contact["properties"], "firstname"|"lastname"|"email">
+  | Pick<Owner, "firstName"|"lastName"|"email">;
 
 const getFullName = (user?: User): string|undefined => {
     if (!user) {
